Only allow a single ondone hook
The original idea was that you could call when_done(..) multiple times
to run different hooks on completion, but that turned out to be
inconvenient for the common case of a single hook, cause we always want
to clear the old hook before registering a new one. Basically, the idea
was half baked, and it's going to get in the way of the next feature.

We can always add that functionality back in later if we want it.